+++ src/Light/SpotLight.h
@@ -8,9 +8,9 @@ class SpotLight : public Light {
 private:
   Point3 origin;
   Vec3 direction;
-  double angle;
+  float angle;
 public:
-  SpotLight(Matrix4 &xform, Color &c, Point3 &o, Vec3 d, double a) : 
+  SpotLight(Matrix4 &xform, Color &c, Point3 &o, Vec3 d, float a) : 
     Light(c), origin(xform.TransformPoint(o)), direction(-xform.TransformVector(d)), angle(a) {}
   Vec3 getDirection(Point3 &p);
   Color diffuseColor (Material* material, Vec3 &n, Vec3 &l);
